Enum unshare::Error

pub enum Error {
    InvalidPath,
    UnknownError,
    CreatePipe(i32),
    Fork(i32),
    Exec(i32),
    Chdir(i32),
    ParentDeathSignal(i32),
    PipeError(i32),
    WaitError(i32),
    StdioError(i32),
    SetUser(i32),
    ChangeRoot(i32),
    SetIdMap(i32),
}

Variants

InvalidPath

Invalid path somewhere when running command. Presumably has embedded nulls.

Frankly, this error should not happen when running process. We just keep it here in case nix returns this error, which should not happen.

UnknownError

Some invalid error code received from child application

CreatePipe

Error happened when we were trying to create pipe. The pipes used for two purposes: (a) for the process's stdio (Stdio::pipe() or Stdio::null()), (b) internally to wake up child process and return error back to the parent.

Fork

Error when forking/cloning process

Exec

Error when running execve() systemcall

Chdir

Error when setting working directory specified by user

ParentDeathSignal

Unable to set death signal (probably signal number invalid)

PipeError

Error reading/writing through one of the two signal pipes

WaitError

Error waiting for process (for some functions only, for example Command::status()). It probably means someone already waited for the process, for example it might be other thread, or signal handler.

StdioError

Error setting up stdio for process

SetUser

Could not set supplementary groups, group id or user id for the process

ChangeRoot

Error changing root, it explains chroot, pivot_root system calls and setting working directory inside new root. Also includes unmounting old file system for pivot_root case.

SetIdMap

Error setting uid or gid map. May be either problem running newuidmap/newgidmap command or writing the mapping file directly

Methods

impl Error

fn raw_os_error(&self) -> Option<i32>

Similarly to io::Error returns bare error code