Enum unshare::ExitStatus

pub enum ExitStatus {
    Exited(i8),
    Signaled(Signal, bool),
}

The exit status of a process

Returned either by reap_zombies() or by child_events() or by Child::wait()

Variants
Exited(i8)

Process exited normally with some exit code

Signaled(Signal, bool)

Process was killed by a signal (bool flag is true when core is dumped)

Methods

impl ExitStatus

fn success(&self) -> bool

Returns true if this exit status means successful exit

fn code(&self) -> Option<i32>

Returns exit code if the process has exited normally

fn signal(&self) -> Option<i32>

Returns signal number if he process was killed by signal