Enum unshare::Stdio

pub enum Stdio {
    Pipe,
    Inherit,
    Null,
    Fd(Closing),
}

An enumeration that is used to configure stdio file descritors

The enumeration members might be non-stable, it's better to use one of the constructors to create an instance

Variants

Pipe

Inherit

Null

Fd

Methods

impl Stdio

fn piped() -> Stdio

Pipe is created for child process

fn inherit() -> Stdio

The child inherits file descriptor from the parent process

fn null() -> Stdio

Stream is attached to /dev/null

fn to_fd(self, write: bool) -> Fd

Converts stdio definition to file descriptor definition (mostly needed internally)