Enum unshare::Stdio

pub enum Stdio {
    Pipe,
    Inherit,
    Null,
    Fd(Closing),
}

An enumeration that is used to configure stdio file descritors

The enumeration members might be non-stable, it's better to use one of the constructors to create an instance

Variants
Pipe

This fd will use pipe to/from the appliation

Inherit

This fd will be inherited from the parent application

Null

This fd will open /dev/null in read or write mode

Fd(Closing)

This is fd passed by application (and closed by unshare)

Methods

impl Stdio

fn piped() -> Stdio

Pipe is created for child process

fn inherit() -> Stdio

The child inherits file descriptor from the parent process

fn null() -> Stdio

Stream is attached to /dev/null

fn to_fd(self, write: bool) -> Fd

Converts stdio definition to file descriptor definition (mostly needed internally)

fn dup_file<F: AsRawFd>(file: &F) -> Result<Stdio>

A simpler helper method for from_raw_fd, that does dup of file descriptor, so is actually safe to use (but can fail)

fn from_file<F: IntoRawFd>(file: F) -> Stdio

A simpler helper method for from_raw_fd, that consumes file

Note: we assume that file descriptor already has the CLOEXEC flag. This is by default for all files opened by rust.